united all assignment tests to one class
