--- src/test/java/app/dao/impl/assignment/DeleteTests.java
@@ -1,29 +1,0 @@
-package app.dao.impl.assignment;
-
-import app.exceptions.EntityNotFoundException;
-import org.dbunit.Assertion;
-import org.dbunit.dataset.IDataSet;
-import org.dbunit.dataset.ITable;
-import org.dbunit.dataset.xml.FlatXmlDataSetBuilder;
-import org.junit.Test;
-
-public class DeleteTests extends AssignmentDaoTestsInitiator {
-
-    @Test
-    public void deleteById() throws Exception {
-        assignmentDao.deleteById(2);
-        IDataSet expectedDataSet = new FlatXmlDataSetBuilder()
-                .build(getClass()
-                        .getClassLoader()
-                        .getResourceAsStream("assignment/deleteDataSets/delete-dataset.xml"));
-        ITable expectedTable = expectedDataSet.getTable(assignmentTable);
-        IDataSet actualDataSet = getMySqlConnection().createDataSet();
-        ITable actualTable = actualDataSet.getTable(assignmentTable);
-        Assertion.assertEquals(expectedTable, actualTable);
-    }
-
-    @Test(expected = EntityNotFoundException.class)
-    public void deleteWithNonExistsPrimaryKey() {
-        assignmentDao.deleteById(4);
-    }
-}
